requirements.txt
habitat-sim==0.3.*
habitat-lab==0.3.*
numpy
